autotest_db/requirements.txt
mysql-connector-python>=8
pandas>=0.25